First implementation of automatic differentiation
Autodiff has been implemented in reverse-mode only. As of now, the four arithmetic
operators +, -, *, / are implemented on scalar values. The
implementation is based on a Wengert list / tape.

A few tests have been written, and have highlighted an issue
with floating point errors. One of the test doesn't pass because of
those rounding errors that keep getting bigger and bigger.

Also, the makefile has been slightly updated to correctly link the .so
file to the test/benchmarks/examples.