Make autodiff classes generic (fixes #2)
Since other classes depend on ts::Var, they have also been made generic
(a solution to make only ts::Var templated would have been to create a
ts::VarBase class, to use pointers of this class as members of the other
classes, and to dynamically cast them to ts::Var<T>*). This way, we can
use our autodiff implementation with any floating point data type that
we may want to support (support has to be added manually for now, since
template classes/functions definitions that are separed from their declaration are not compiled : we would need to put them in the same header file otherwise).

A ts::NewVar function has also been created. This way, if we want to
avoid using the template syntax while creating ts::Var or ts::Gradient
instances, and write our desired data type only
once (while creating our ts::WengertList), we can use this function along with the
auto keyword (see updated tests). This is still optional and has the
inconvenient of making it harder to figure out the type of our variables.

In the future, it might be interesting to create benchmarks to compare
the performances/accuracies of different floating point data types, and
add support for more of them.