Add matrix product
A function to compute a classic matrix product has been added. Since
thus operation is not element-wise, there are now 2 possibilities :

-All operations are element-wise, we allow to compute the gradient with
 respect to a non scalar output like before

-We have at least one non-element-wise operation, so the output variable
has to be a scalar, and the derivative with respect to a matrix will still be a matrix
with the same dimensions

The gradient of the matrix product still needs to be tested, next commit will probably add a
test, as well the norm and sigmoid functions to simulate a simple feedforward NN.

Overloaded operators and tensors functions have been moved to a new
file.