Complete minimal implementation of tensor autodiff
Added the element-wise sigmoid and squared norm functions on tensors.
To validate our implementation, a test simulating a minimal feedforward NN has been written. We compute the forward pass,the value of the cost function, as well as its gradient. We compare the
obtained values with the pre-computed results (using PyTorch Autograd).
This example uses everything we've developed so far, including
matrix-matrix product differentiation, and is very similar to our real
use case.
The current state is considered sufficient to start working on NN, and
to merge with master.