Prepare autodiff engine for models and optimizations implementation
A pointer to the ts::Tensor has been added to ts::InputNode. This way,
it will be accessible in the optimization phase so its value can be
adjusted.

A method to clean ts::WengertList has been added. This remove all
non-input nodes, so we can run a model multiple times during training
(or even prediction) phase.
NOTE: Maybe a better optimization in the future would be to keep the
dependency graph once built, and simply update the values ?