Add Model and Optimizer class (not tested yet!)
The ts::Model class is supposed to encapsulate any mathematical model
that can be optimized. The library will propose basic models (especially
different kinds of neural networks), but a user can define its own.

The ts::Optimizer is supposed to be a base class to implement different
optimization algorithms that will modify values of ts::Tensors contained
in a ts::Model. In order to achieve this, inputNodes can now have a
pointer to their base tensor, so it's possible to access their values in
an optimizer.

For now one instance for each of these virtual are being implemented :
- ts::MultiLayerPerceptron for ts::Model
- ts::GradientDescentOptimizer for ts::Optimizer

In the near future, these classes will be corrected/completed and tests
will be written for them.