Make models and gradient descent functionnal
Polynom and multilayer perceptron are now functionnal and tested.

Nodes pointing to optimizable tensors now include a pointer to the
tensor. The SGD optimizer should be working, but writing full tests will
be required.

In the future, those tests will be written, and the run function of
optimizers may return the different losses in order to review the
training results and ensure they are good.