Write test for gradient descent optimizer
The ts::Model::run() method now returns a 3D array containing the cost
of all runs. This way, data can be analyzed at the end of the training
phase.

For instance, this data is used in the new test validating our gradient
implementation on a polynom.

This version of the code seems functionnal an can further be tested on a
more complex example, such as solving MNIST