Add code to read MNIST binary files
This is the first version of the code that generates a vector of
ts::TrainingData. It will probably be put in a function in the next
commit to be called on both training and testing data.
It can also be optimized by not reading the entire file but only the
beginning if we're ok with always using the first digits of the files.

Finally, minor changes have been made to other files, including :
- Renaming the "test" directory to "tests"
- Correcting typos
- Adding a few comments
- Updating the README