First version of MNIST example
The code to read MNIST data files has been corrected/improved, and the
example has been completed. A MLP model is created, trained and finally
tested. The user can then visualize the results through the terminal.

A few changes has been made to the library code. Especially, the
optimization must now be toggled manually for an entire model. Before
that, tensor were made optimizabe in the constructor, which could cause
issues if the method was called in a loop for instance. This should
avoid the issue, and makes it possible to allow/disallow optimization
and even to define other optimizations methods for one model.

Finally a bug has been found when differentiating a product of two
matrices with the same size. An issue will be created on the repo to fix
it as soon as possible.