Complete Adam optimizer
Adam optimizer is now functionnal and has been tested on the MNIST
example with good results. It still needs a dedicated test for
completeness sake, and it might be possible to optimize it in the future.

Finally, the code relies on the original 2015 paper "Adam: A method for
stochastic optimization" by Diederik P. Kingma and Jimmy Lei Ba.
Complete explanations are given about the algorithm, which will make the
last two commits much easier to understand.