Bugfix for convolution ops derivatives
The calculation of convolution and max pooling partial derivatives had
issues that have been corrected. Commit also fixes/improves other aspects
of our CNN model.

However, while testing the CIFAR example, results don't seem to improve
after training the model (at least not significantly). This might be an
issue with the way our example is implemented, with the CNN itself or
with derivatives computations. This should be fixed as soon as possible
in future commits to validate our CNN implementation.