Add channels to convolution layers
Previous version lacked the ability to define multiple channels for one
convolution layer. It is now possible to define this number using the
same constructor syntax for the CNN model.

CIFAR example still performs badly. It could be either because of
derivatives errors in convolution operations (might be a good idea
to use a librady such as Pytorch/Autograd to test them), because
of the way forward pass is calculated in the CNN, or just because
the setup for CIFAR is wrong.