Add channel splting before first convolution layer
Previous version forced us to start convolutions with a single channel
input. Now, the input can be splitted either vertivally or horizontally
in order to create multiple channels. We might want to add a dedicated
operation with its derivative because as of now, we just create new
tensors from the input value, and can't get the derivative with respect
from this input (even though we probably won't need it).