Start OpenMP parallelization
For now, the main benefit is the parallelization of arrayConv function
(which is the one costing the more computations in our CNN). Another way
of speeding this up would be to implement more optimized algorithms such
as the one described in :
https://blog.usejournal.com/understanding-winograd-fast-convolution-a75458744ff

It might also be possible to parallelize optimization algorithms, but
would require a copy of the wList/entire model for every thread.